pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "openblend-public"
version = "0.1.0"
description = "🍸 Pre-trained blended LLM - Outperforms any single model by combining your premium model with free models via ELO-ranked Steer-then-Verify"
readme = "README.md"
requires-python = ">=3.10"
license = "MIT"
authors = [{name = "OpenBlend Maintainers"}]
keywords = ["llm", "blend", "openai-compatible", "elo-ranking", "multi-model"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "typer[all]>=0.9",
    "rich>=13.0",
    "pyyaml>=6.0",
    "python-dotenv>=1.0",
    "fastapi>=0.100",
    "uvicorn[standard]>=0.23",
    "pydantic>=2.0",
    "aiosqlite>=0.19",
    "openai>=1.0",
    "httpx>=0.27.0",
]

[project.scripts]
openblend = "openblend_public.cli:app"

[project.urls]
Homepage = "https://github.com/jajabong/openblend-public"
Repository = "https://github.com/jajabong/openblend-public"
Issues = "https://github.com/jajabong/openblend-public/issues"

[tool.ruff]
line-length = 100
target-version = "py310"
